+++ include/robot.h
@@ -68,22 +68,22 @@ class Robot {
 
     void moveArmTo(double degr, double speed);
 
-    void smartDrive(float distInches, float speed, directionType left, directionType right, int timeout, float slowDownInches, 
+    void smartDrive(float distInches, float speed, directionType left, directionType right, float timeout, float slowDownInches, 
                     float turnPercent, bool stopAfter, std::function<bool(void)> func);
-    void driveTurn(float degrees, float speed, bool isClockwise, int timeout, float slowDownInches = 10, 
+    void driveTurn(float degrees, float speed, bool isClockwise, float timeout, float slowDownInches = 10, 
                     bool stopAfter = true, std::function<bool(void)> func = {});
-    void driveCurved(float distInches, float speed, directionType dir, int timeout, 
+    void driveCurved(float distInches, float speed, directionType dir, float timeout, 
                       float slowDownInches, float turnPercent, bool stopAfter = true, std::function<bool(void)> func = {});
-    void driveStraight(float distInches, float speed, directionType dir, int timeout, 
+    void driveStraight(float distInches, float speed, directionType dir, float timeout, 
                       float slowDownInches, bool stopAfter = true, std::function<bool(void)> func = {});
-    void driveStraightTimed(float speed, directionType dir, int timeMs, bool stopAfter = true, std::function<bool(void)> func = {});
+    void driveStraightTimed(float speed, directionType dir, float timeMs, bool stopAfter = true, std::function<bool(void)> func = {});
 
     void goForwardVision(Goal goal, float speed, directionType dir, float maximumDistance, int timeout, 
                         digital_in* limitSwitch = nullptr, std::function<bool(void)> func = {});
     void alignToGoalVision(Goal goal, bool clockwise, directionType cameraDirection, int timeout);
     void updateCamera(Goal goal);
 
-    void driveStraightGyro(float distInches, float speed, directionType dir, int timeout, float slowDownInches,
+    void driveStraightGyro(float distInches, float speed, directionType dir, float timeout, float slowDownInches,
                             std::function<bool(void)> func = {});
     void turnToAngleGyro(bool clockwise, float angleDegrees, float maxSpeed, int startSlowDownDegrees,
                         int timeout, std::function<bool(void)> func = {});
